Removing string_decode_buffer and fixed potential unsafe access with table lookups (#1168)
* Removing string_decode_buffer

* Removing branches in string writing

* Convert to unsigned uint8_t for table lookups

+++ include/glaze/json/json_format.hpp
@@ -181,7 +181,7 @@ namespace glz::detail
    inline sv read_json_number(auto&& it) noexcept
    {
       auto start = it;
-      while (numeric_table[*it]) {
+      while (numeric_table[uint8_t(*it)]) {
          ++it;
       }
       return {start, size_t(it - start)};
